# [converter, exception, sql]

+++ src/main/java/br/com/hoffmann/service/generic/impl/GenericCrudServiceImpl.java
@@ -6,6 +6,7 @@ import br.com.hoffmann.service.exception.EntityNotFoundException;
 import br.com.hoffmann.service.exception.ServiceException;
 import br.com.hoffmann.service.generic.GenericCrudService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 
 
 public class GenericCrudServiceImpl <T extends BaseEntity> implements GenericCrudService<T> {
@@ -15,6 +16,7 @@ public class GenericCrudServiceImpl <T extends BaseEntity> implements GenericCru
    protected GenericCrudRepository<T> repository;
 
     @Override
+    @Transactional
     public T save(T entity) {
         try {
            return repository.save(entity);
@@ -24,12 +26,19 @@ public class GenericCrudServiceImpl <T extends BaseEntity> implements GenericCru
     }
 
     @Override
+    @Transactional
     public void deleteById(Long id) {
         try {
+            if(!repository.existsById(id)){
+                throw new EntityNotFoundException("Id not found "+id);
+            }
             repository.deleteById(id);
+        }catch (EntityNotFoundException ex){
+            throw new EntityNotFoundException(ex);
         }catch (Exception ex){
-            throw new ServiceException(ex);
+            throw new ServiceException("Não foi possível remover objeto com id "+id, ex);
         }
+
     }
 
     @Override
